refactor: Improve file handling (std::filesystem::path, slightly more efficiend Shader::readFile(), smaller shader compile message buffers)

+++ src/renderer.h
@@ -34,7 +34,7 @@ public:
     // Screen update and buffer swap is responsibility of window
 
 private:
-    enum class ShaderInstance : size_t
+    enum class ShaderInstance : uint8_t
     {
         ModelShader,
         SkyboxShader,
